refactor(ops): ✨ consolidate all ops into main.go
The seven ops_*.go files are merged into main.go. All implementations
live in one place alongside the CLI dispatch that calls them. The
separate files were a holdover from the per-op Rust module structure;
in Go they add navigation indirection without meaningful benefit.

default.nix srcs list is updated to remove the now-deleted files.
